+++ flexmeasures/utils/time_utils.py
@@ -279,15 +279,6 @@ def forecast_horizons_for(resolution: str | timedelta) -> list[str] | list[timed
         return horizons
 
 
-def supported_horizons() -> list[timedelta]:
-    return [
-        timedelta(hours=1),
-        timedelta(hours=6),
-        timedelta(hours=24),
-        timedelta(hours=48),
-    ]
-
-
 def timedelta_to_pandas_freq_str(resolution: timedelta) -> str:
     return to_offset(resolution).freqstr